gui: Restore screenshot capability
Hijacking the power key event handling killed fb2png's screenshot
feature. Avoid processing screen blank toggle if volume down key is
either pressed or was recently pressed (< 3 s).

Note that NotifyKey in GUIActions is not being used because there
seem to be some issues with 1) two-key screenshot action causes
power key UP to be gobbled so single key action cannot be created,
and 2) it does not respond when device screen is blanked (presumably
thanks to the lock page).

Change-Id: Ia4d78e3ddc59cb51e748876155c1543ba3671452