libtar: Cleanup, secure, and extend numeric fields
Cleanup:

libtar emits numeric field data without leading zeros and terminates
fields with a space and NUL, except for the mtime and size fields which
are only terminated by a space (when sufficiently large).

Fix this to emit numeric fields in the standard and recommended way,
which is padded with leading zeros and always terminated by a NUL.

Secure:

libtar does not supply a length parameter to oct_to_int(), and relies
on the data stream to contain invalid octal values that will terminate
sscanf() parsing.  A specially crafted tar file could potentially
exploit this and cause libtar to overrun its input buffer.

Fix this by supplying a length field to oct_to_int() and copying the
data into a temporary buffer that is always NUL terminated.

Extend:

libtar follows the original tar file format, which allots 12 bytes for
the size field, the last of which should be a NUL (or space) for
compatibility reasons.  This results in a maximum file size of 8gb.

Many (most?) modern tar implementations support a binary extension for
the size and certain other fields to remedy this situation.  This is
indicated by setting the high bit in the first byte and placing the
"base-256" numerical value (eg. big endian binary representation) in
the field.

Use the binary extension for the mtime and size fields.

Change-Id: If1fc3dd191143db245539386a64045eff7f12869