Restore ability to add /system_image fstab lines
TWRP allows for the system and vendor partitions to be repeated in
fstab as long as they have distinct mount points:
  /system_image and /vendor_image
Since we currently skip over duplicate block-devices in fstab, we
weren't honoring this feature. Frankly, it's terribly implemented, but
is technically a regression, so I'll restore the ability by instead
skipping over partitions with duplicate mount-points. The feature
needs to be rewritten to accept image backup via a twrp flag, without
an additional line in fstab.

Change-Id: I9c9dd50067f474207eae3a42f2af763de4cf1650